Create type-guards module and expose it
