+++ packages/core/src/new-api/internal/isFutureType.ts
@@ -1,0 +1,8 @@
+import { FutureType } from "../types/module";
+
+export function isFutureType(potential: unknown): potential is FutureType {
+  return (
+    typeof potential === "string" &&
+    (FutureType as any)[potential] !== undefined
+  );
+}
