+++ packages/core/src/new-api/stored-deployment-serializer.ts
@@ -23,7 +23,7 @@ import {
   isContractFuture,
   isFuture,
   isRuntimeValue,
-} from "./internal/utils";
+} from "./type-guards";
 import {
   AccountRuntimeValue,
   AddressResolvableFuture,
