+++ packages/core/src/ui-helpers.ts
@@ -4,3 +4,4 @@ export {
 } from "./new-api/stored-deployment-serializer";
 export * from "./new-api/types/module";
 export * from "./new-api/types/serialized-deployment";
+export * from "./new-api/type-guards";
